requirements.txt
# ---- API Server ----
fastapi==0.110.0
uvicorn==0.29.0

# ---- LLM (LiteLLM via OpenAI-compatible SDK) ----
openai==1.14.3

# ---- LLM (Google AI/Gemini SDK) ----
google-generativeai==0.8.3

# ---- Google Analytics 4 ----
google-analytics-data==0.18.6
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# ---- Google Sheets (SEO Agent) ----
gspread==6.1.2
google-api-python-client==2.125.0

# ---- Data Processing ----
pandas==2.2.2
numpy==1.26.4

# ---- Validation & Utilities ----
pydantic==2.6.4
python-dateutil==2.9.0.post0

# ---- HTTP & Retry Handling ----
httpx==0.27.0
tenacity==8.2.3

# ---- Environment ----
python-dotenv==1.0.1

